[#694] change data type for logginginfo for transaction

+++ thrift/src/main/thrift/Trace.thrift
@@ -101,7 +101,7 @@ struct TSpan {
   26: optional TIntStringValue exceptionInfo;
   
   30: optional i16 applicationServiceType;
-  31: optional i16 loggingTransactionInfo;
+  31: optional byte loggingTransactionInfo;
 }
 
 struct TSpanChunk {
